pyproject.toml
[project]
name = "gemini-mcp"
version = "0.1.0"
description = "MCP server for image generation and editing with Gemini API and ImageMagick"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Olivier", email = "om@kosmokompany.com" }
]
keywords = ["mcp", "gemini", "image-generation", "imagemagick", "ai"]

dependencies = [
    "mcp[cli]>=1.2.0",
    "google-genai>=1.0.0",
    "wand>=0.6.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]

[project.scripts]
gemini-mcp = "gemini_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gemini_mcp"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
